Change compute implement (#45)
* Deleted Unused folder

I deleted an unused folder in src called "Test" that could lead to confusion. It only contained a duplicate of the importme.txt file.

* Changed the Compute Implement API to "Networker"

I changed the computeImplement API to "Networker", going off of some of the API examples Professor Brainard provided. I established an inputConfiguration, outputConfiguration, and Delimiter field, as well as a default delimiter and constructors for if a delimiter is specified or not.

* Fixed Misnamed File for OutputConfiguration.java

* More mismatched name fixes

Filename:
TestComputeImpl.java => TestNetworker.java

Changed names within this file to reflect new name

* Changed TestComputeImplement to TestNetworker

Changed name as well as created a new test to test the new functionalities of the Networker API